Use separate files for TraceId and SpanId
